Add match by values
It will work similarly to how `for` works: Instead of creating a new
`Statement` in the AST, we refactor it using existing ones. That will
work like:

```
match v {
  1 => {},
  2 => {},
  3 => {},
  * => {},
}
```

Will be translated into:

```
{
  var @match_value = v;
  if (@match_value == 1) {}
  else if (@match_value == 2) {}
  else if (@match_value == 3) {}
  else {}
}
```

Note that the temporary variable, `@match_value` won't polute the user
scope because we don't allow identifiers with `@` for them.