Implement istype
It will return true if the type of the primary values matches the right
hand argument or, if the left hand is an object, if it is a direct
implementation of the class on the right side, or subclass of it, or if
it implements the trait specified.

We will use this later to do the match for types.