Remove unnecessary lexeme
We don't need to go carrying that thing around. It is a very easy win to
just remove it.

Also use &str instead of String for file names.

No need to maintain thousands copies of the same string.

That requires a huge (HUGE) refactor on how the passes work and how we
import staff.