Fix ObjectSet error
When ObjectSet requires to resize the object, it might end up with
a different address and thus being a completely different object

If the object was also a local or a constant, that means having to
update all the references to it. Instead, this solution uses
indirection to avoid creating a new object in the same place and
always referring to the same one.